refactor: dont actually want it as a screen
