+++ app/src/main/java/app/revanced/manager/ui/model/navigation/Nav.kt
@@ -17,9 +17,6 @@ object Dashboard
 object AppSelector
 
 @Serializable
-data class AppliedPatches(val packageName: String)
-
-@Serializable
 data class InstalledApplicationInfo(val packageName: String)
 
 @Serializable
